packages/agent/requirements.txt
pipecat-ai[daily,google,silero]>=0.0.60
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0
